refactor: add redeem endpoint

+++ src/main/java/hoomgroom/product/Promo/controller/PromoController.java
@@ -1,45 +1,70 @@
-package hoomgroom.product.Promo.controller;
+package hoomgroom.product.promo.controller;
 
-import hoomgroom.product.Promo.model.Promo;
-import hoomgroom.product.Promo.service.PromoService;
+import hoomgroom.product.auth.service.JwtServiceImpl;
+import hoomgroom.product.promo.dto.*;
+import hoomgroom.product.promo.model.Promo;
+import hoomgroom.product.promo.service.PromoService;
+import jakarta.servlet.http.HttpServletRequest;
+import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.UUID;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
 
 @RestController
 @RequestMapping("/api/promo")
 @RequiredArgsConstructor
 public class PromoController {
     private final PromoService promoService;
+    private final JwtServiceImpl jwtService;
 
     @GetMapping("/")
-    public ResponseEntity<List<Promo>> getAllPromos() {
-        List<Promo> response = promoService.findAll();
-        return ResponseEntity.ok(response);
+    public CompletableFuture<ResponseEntity<List<Promo>>> getAllPromos() {
+        return promoService.findAll().thenApply(ResponseEntity::ok);
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<?> getPromoById(@PathVariable UUID id) {
-        try {
-            Promo response = promoService.findById(id);
-            return ResponseEntity.ok(response);
-        } catch (NoSuchElementException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Promo not found");
-        }
+    public CompletableFuture<ResponseEntity<PromoResponse>> getPromoById(@PathVariable UUID id) {
+        return promoService.findById(id);
     }
 
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<String> deletePromo(@PathVariable UUID id) {
-        try {
-            promoService.delete(id);
-            return ResponseEntity.ok(String.format("Successfully deleted Promo with id %s", id.toString()));
-        } catch (NoSuchElementException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Promo not found");
+    public CompletableFuture<ResponseEntity<Response>> deletePromo(
+            @NonNull HttpServletRequest request,
+            @PathVariable UUID id
+    ) throws ExecutionException, InterruptedException {
+        boolean isAuthenticated = jwtService.isAuthenticated(request);
+
+        if (!isAuthenticated) {
+            Response response = Response
+                    .builder()
+                    .message("ADMIN Only")
+                    .build();
+            return CompletableFuture.completedFuture(ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(response));
         }
+        return promoService.delete(id);
+    }
+
+    @GetMapping("/redeem/{id}")
+    public CompletableFuture<ResponseEntity<RedeemResponse>> redeemPromo(
+            @NonNull HttpServletRequest request,
+            @PathVariable UUID id,
+            @RequestBody RedeemPromoRequest requestBody
+    ) {
+        boolean isAuthenticated = jwtService.isAuthenticated(request);
+
+        if (!isAuthenticated) {
+            RedeemResponse response = RedeemResponse
+                    .builder()
+                    .message("ADMIN Only")
+                    .build();
+            return CompletableFuture.completedFuture(ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(response));
+        }
+        return promoService.redeem(requestBody.getTransactionId(), id, requestBody.getTotalPrice());
     }
 }
